+++ src/store/actions/actions.ts
@@ -1,4 +1,4 @@
-import { INCREMENT, DECREMENT } from "./types";
+import { INCREMENT, DECREMENT } from "../types";
 
 export const increment = () => ({
   type: INCREMENT,
