+++ src/store/reducers/counter-reducer/counter.reducer.ts
@@ -1,6 +1,4 @@
-// reducers/counterReducer.ts
-
-import { INCREMENT, DECREMENT } from "../../actions/types";
+import { INCREMENT, DECREMENT } from "../../types";
 
 const initialState = {
   count: 0,
